size_t for size variables

+++ structuresModulesFiles/phoneBook/phoneBook/phoneBook.h
@@ -5,7 +5,7 @@
 #include <wchar.h>
 
 #define PHONEBOOK_MAX_ENTRIES (size_t)100
-#define PHONEBOOK_NUMBER_LENGTH_MAX 15
+#define PHONEBOOK_NUMBER_LENGTH_MAX (size_t)15
 #define PHONEBOOK_NAME_LENGTH_MAX 50 * sizeof(wchar_t)
 
 typedef struct phoneBookEntry {
@@ -15,7 +15,7 @@ typedef struct phoneBookEntry {
 
 struct phoneBook {
     phoneBookEntry entries[PHONEBOOK_MAX_ENTRIES];
-    int size;
+    size_t size;
     bool saved;
 };
